Rewrite get_my_public_url so that it doesn’t depend on `request` variable anymore.
